Separacion de carrito en app
