+++ vitrina-project/frontend/src/components/Carrito/useCart.js
@@ -1,0 +1,54 @@
+// src/hooks/useCart.js
+import { useState } from 'react';
+
+export const useCart = () => {
+  const [cart, setCart] = useState([]);
+  const [showConfirmation, setShowConfirmation] = useState(false);
+
+  const handleAddToCart = (product) => {
+    setCart(prevCart => {
+      const existingProduct = prevCart.find(item => item.id === product.id);
+      if (existingProduct) {
+        return prevCart.map(item =>
+          item.id === product.id ? { ...item, quantity: item.quantity + 1 } : item
+        );
+      } else {
+        return [...prevCart, { ...product, quantity: 1 }];
+      }
+    });
+
+    setShowConfirmation(true); // Mostrar el mensaje de confirmación
+    setTimeout(() => {
+      setShowConfirmation(false); // Ocultar el mensaje después de 3 segundos
+    }, 3000);
+  };
+
+  const handleIncrement = (id) => {
+    setCart(prevCart =>
+      prevCart.map(item =>
+        item.id === id ? { ...item, quantity: item.quantity + 1 } : item
+      )
+    );
+  };
+
+  const handleDecrement = (id) => {
+    setCart(prevCart =>
+      prevCart
+        .map(item =>
+          item.id === id ? { ...item, quantity: item.quantity - 1 } : item
+        )
+        .filter(item => item.quantity > 0)
+    );
+  };
+
+  const totalItemsInCart = cart.reduce((total, item) => total + item.quantity, 0);
+
+  return {
+    cart,
+    showConfirmation,
+    handleAddToCart,
+    handleIncrement,
+    handleDecrement,
+    totalItemsInCart
+  };
+};
